requirements.txt
python-Levenshtein
pydub
numpy
nltk==3.2.5
resampy
pandas==0.23.0
six
mpi4py
librosa
python_speech_features
matplotlib
joblib
sentencepiece
sacrebleu
h5py
tqdm
